# server/openapi.yaml
openapi: 3.0.3
info:
  title: ActivityOS API
  version: 0.31.0
  description: |
    Express + Firestore backend for ActivityOS — a **multi-tenant** platform
    for children's activity providers.

    ## v0.31.0 — Cross-provider marketplace (opt-in)

    The parent Browse feed becomes a real marketplace instead of being scoped
    (client-side) to providers a family had already booked.
    - **`GET /api/listings`** (no `?tenantId=`, no `?mine=1`) now returns the
      live/public/unarchived listings of every provider **that has opted in**
      (`settings.marketplaceListed === true`) — **plus**, for a signed-in
      parent, their own providers (anyone they've booked) regardless of opt-in.
    - A provider's **own storefront** (`GET /api/listings?tenantId=`) is
      unchanged — their public listings always show there whether or not
      they're in the marketplace.
    - New tenant setting **`marketplaceListed`** (in the library `settings`
      bag), default **false** — a provider opts in from Setup → Bookings →
      Marketplace. Opt-in is a deliberate choice; nobody is publicly listed
      without it.
    (Frontend: BrowseApp drops the old single-provider scoping and shows the
    server feed as-is.)

    ## v0.30.0 — Account, Data & privacy, parent Accidents

    Clearing the shared/parent mock pages.
    - **`/api/account`** (every role) — the signed-in user's profile. `GET`
      returns `{email, name, phone, marketingConsent, role, tenantId}` (identity
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - firebaseIdToken: []
paths:
...
  /api/meal-orders/{id}/pay:
    post:
      tags: [meal-orders]
      summary: Record payment on an order (operators) → Paid
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: The order (pay → Paid) }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: The order was cancelled }
...
  /api/payments/connect:
    post:
      tags: [payments]
      summary: Connect (or resume) the tenant's Stripe Express account
      responses:
        "200":
          description: The Stripe-hosted onboarding URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }
                  accountId: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503":
          description: Payments not configured on the server
  /api/payments/status:
    get:
      tags: [payments]
      summary: How connected the tenant's Stripe account is
      responses:
        "200":
          description: Connection state
          content:
            application/json:
              schema:
                type: object
                properties:
                  connected: { type: boolean }
                  accountId: { type: string }
                  chargesEnabled: { type: boolean }
                  detailsSubmitted: { type: boolean }
                  payoutsEnabled: { type: boolean }
                  platformFallback: { type: boolean, description: DEV flag — see the v0.9.0 notes }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/payments:
    get:
      tags: [payments]
      summary: The tenant's payment & refund records (oversight)
      parameters:
        - name: tenantId
          in: query
          required: false
          description: Platform accounts only
          schema: { type: string }
      responses:
        "200":
          description: Records, newest first (refunds have type "refund")
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/payments/checkout:
    post:
      tags: [payments]
      summary: Start paying for bookings (parent)
      description: |
        One PaymentIntent for the given bookings (all one provider, all
        payable — status `Confirmed` or `pay: "Invoice sent"`), created as a
        direct charge on the provider's connected account. Initialise
        Stripe.js with the returned `stripeAccount` before mounting the
        Payment Element.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refs]
              properties:
                refs:
                  type: array
                  items: { type: string }
      responses:
        "201":
          description: Client secret for the Payment Element
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentId: { type: string }
                  clientSecret: { type: string }
                  stripeAccount: { type: string, nullable: true }
                  amount: { type: number }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Not payable (already paid, not confirmed, or the provider hasn't finished Stripe onboarding)
  /api/payments/checkout/{id}/confirm:
    post:
      tags: [payments]
      summary: Verify the payment with Stripe and mark the bookings Paid
      description: Idempotent; only the paying family's account can call it.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: "{status, paid} — bookings flip to Paid when succeeded"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token (see the Authentication section above).
...
  responses:
    BadRequest:
      description: Validation failed (`error` is a message or zod issue list)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing/invalid Firebase ID token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The account's role/tenant doesn't allow this
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found (or outside the caller's tenant scope)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          description: Human-readable message, or an array of zod issues on 400.
          oneOf:
            - type: string
            - type: array
              items: { type: object }
...